# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: SkillSnap - AI-powered local services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /providers:
    get:
      operationId: getProviders
      tags: [providers]
      summary: Get list of providers
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking from a service request and matched provider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get user's bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [consumer, provider]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
  /payments:
    get:
      operationId: listMyPayments
      tags: [payments]
      summary: List payments for the authenticated consumer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Bookings and revenue by month (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Time series for charts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalyticsResponse"
  /admin/category-stats:
    get:
      operationId: getAdminCategoryStats
      tags: [admin]
      summary: Per-category provider counts, bookings, avg rating (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCategoryStatRow"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: Paginated users with optional role filter (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: role
          in: query
          schema:
            type: string
            enum: [consumer, provider, admin]
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/providers:
    get:
      operationId: listAdminProviders
      tags: [admin]
      summary: Paginated provider profiles with user summary (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: verificationStatus
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected]
      responses:
        "200":
          description: Paginated providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProvidersPage"
  /admin/providers/verification:
    patch:
      operationId: patchAdminProviderVerification
      tags: [admin]
      summary: Approve or reject a provider profile (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProviderVerificationRequest"
      responses:
        "200":
          description: Updated provider with user summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProviderWithUser"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [admin]
      summary: Global booking list with joins (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingsPage"
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: Payments with consumer/provider names (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPaymentsPage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [consumer, provider]
      required:
        - fullName
        - email
        - password
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [consumer, provider, admin]
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - icon
        - isActive
    Provider:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        bio:
          type: string
        yearsExperience:
          type: number
        avgRating:
          type: number
        totalJobs:
          type: number
        completionRate:
          type: number
        acceptanceRate:
          type: number
        serviceRadiusKm:
          type: number
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        basePrice:
          type: number
        priceType:
          type: string
          enum: [hourly, fixed]
        distance:
          type: number
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - userId
        - businessName
        - avgRating
        - totalJobs
        - completionRate
        - acceptanceRate
        - serviceRadiusKm
        - verificationStatus
        - categories
        - basePrice
        - priceType
...
    CreateBookingInput:
      type: object
      properties:
        serviceRequestId:
          type: string
        providerId:
          type: string
        scheduledAt:
          type: string
      required:
        - serviceRequestId
        - providerId
        - scheduledAt
    Booking:
      type: object
      properties:
        id:
          type: string
        serviceRequestId:
          type: string
        consumerId:
          type: string
        providerId:
          type: string
        status:
          type: string
          enum: [requested, matched, accepted, on_the_way, arrived, in_progress, completed, cancelled]
        scheduledAt:
          type: string
        acceptedAt:
          type: string
          nullable: true
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        cancelledAt:
          type: string
          nullable: true
        finalPrice:
          type: number
          nullable: true
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        createdAt:
          type: string
      required:
        - id
        - serviceRequestId
        - consumerId
        - providerId
        - status
        - scheduledAt
        - paymentStatus
        - createdAt
...
    PaymentRecord:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        provider:
          type: string
        externalPaymentId:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        paidAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - bookingId
        - amount
        - currency
        - status
        - createdAt
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: number
        totalProviders:
          type: number
        totalBookings:
          type: number
        activeBookings:
          type: number
        completedBookings:
          type: number
        totalRevenue:
          type: number
        pendingVerifications:
          type: number
        avgRating:
          type: number
        completionRate:
          type: number
        openDisputes:
          type: number
      required:
        - totalUsers
        - totalProviders
        - totalBookings
        - activeBookings
        - completedBookings
        - totalRevenue
        - pendingVerifications
        - avgRating
        - completionRate
        - openDisputes
    AdminAnalyticsResponse:
      type: object
      properties:
        series:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              bookings:
                type: number
              revenue:
                type: number
            required: [month, bookings, revenue]
      required:
        - series
    AdminCategoryStatRow:
      type: object
      properties:
        categoryId:
          type: string
        name:
          type: string
        providerCount:
          type: number
        bookingCount:
          type: number
        avgProviderRating:
          type: number
      required:
        - categoryId
        - name
        - providerCount
        - bookingCount
        - avgProviderRating
    AdminUsersPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: number
      required:
        - items
        - total
    AdminProvidersPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: number
      required:
        - items
        - total
    AdminProviderVerificationRequest:
      type: object
      required:
        - providerId
        - verificationStatus
      properties:
        providerId:
          type: string
          description: Provider profile id (MongoDB _id)
        verificationStatus:
          type: string
          enum: [verified, rejected]
        reason:
          type: string
          description: Optional note shown to provider when rejected
    AdminUserSummary:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - fullName
        - email
    AdminProviderWithUser:
      type: object
      description: Provider profile row with owner summary (admin PATCH/list item shape)
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        bio:
          type: string
        yearsExperience:
          type: number
        avgRating:
          type: number
        totalJobs:
          type: number
        completionRate:
          type: number
        acceptanceRate:
          type: number
        serviceRadiusKm:
          type: number
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/AdminUserSummary"
      required:
        - id
        - userId
        - businessName
        - verificationStatus
        - createdAt
        - updatedAt
        - user
    AdminBookingsPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: number
      required:
        - items
        - total
    AdminPaymentsPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: number
      required:
        - items
        - total